+++ app/src/main/java/com/example/messapp/Adapter/UserAdapter.java
@@ -1,0 +1,62 @@
+package com.example.messapp.Adapter;
+
+import android.content.Context;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.ImageView;
+import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.messapp.R;
+import com.example.messapp.User;
+
+import java.util.List;
+
+public class UserAdapter extends RecyclerView.Adapter<UserAdapter.ViewHolder> {
+
+    private Context mContext;
+    private List<User> mUsers;
+
+    public UserAdapter (Context mContext, List<User> mUsers){
+        this.mUsers = mUsers;
+        this.mContext = mContext;
+
+    }
+
+    @NonNull
+    @Override
+    public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
+        View view = LayoutInflater.from(Context).inflate(R.layout.user_item, parent, false);
+        return null;
+    }
+
+    @Override
+    public void onBindViewHolder(@NonNull ViewHolder holder, int position) {
+
+    }
+
+    @Override
+    public int getItemCount() {
+        return mUsers.size();
+    }
+
+    public class ViewHolder extends RecyclerView.ViewHolder {
+        public TextView username;
+        public ImageView imageView;
+
+        public ViewHolder(View itemView){
+            super(itemView);
+
+
+            username = itemView.findViewById(R.id.username);
+            imageView = itemView.findViewById(R.id.user_image);
+        }
+
+    }
+
+
+
+}
