+++ app/src/main/java/com/example/messapp/Fragment/MessageFragment.java
@@ -1,0 +1,23 @@
+package com.example.messapp.Fragment;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.example.messapp.R;
+
+
+public class MessageFragment extends Fragment {
+
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_message, container, false);
+    }
+}
