+++ app/src/main/java/com/example/messapp/Fragment/UsersFragment.java
@@ -1,0 +1,22 @@
+package com.example.messapp.Fragment;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
+
+import com.example.messapp.R;
+
+
+public class UsersFragment extends Fragment {
+
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_users, container, false);
+    }
+}
