+++ app/src/main/res/layout/fragment_message.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorAccent"
-    tools:context=".MessageFragment">
+    tools:context=".Fragment.MessageFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
