+++ app/src/main/res/layout/fragment_profile.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/fragment3"
-    tools:context=".MessageFragment">
+    tools:context=".Fragment.ProfileFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
