+++ app/src/main/res/layout/fragment_users.xml
@@ -1,18 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/fragment2"
-    tools:context=".MessageFragment">
+    tools:context=".Fragment.UsersFragment">
 
     <!-- TODO: Update blank fragment layout -->
-    <TextView
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/recycle_view"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_margin="20dp"
-        android:text="Fragment 2"
-        android:textColor="@color/white"
-        android:textSize="20dp" />
+        android:layout_height="match_parent" />
 
-</FrameLayout>
+</RelativeLayout>
