+++ app/src/main/res/layout/user_item.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:padding="10dp"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content">
+
+
+    <com.mikhaellopez.circularimageview.CircularImageView
+        android:layout_width="70dp"
+        android:layout_height="70dp"
+        android:src="@drawable/image"
+        android:id="@+id/user_image"
+        app:civ_border="true" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="username"
+        android:id="@+id/user"
+        android:layout_toRightOf="@id/user_image"
+        android:layout_toEndOf="@id/user_image"
+        android:layout_marginLeft="15dp"
+        android:layout_centerVertical="true"
+        android:textSize="20dp"
+        />
+
+</RelativeLayout>
